+++ strings/strings.c
@@ -1,48 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-
-/*
-    Given a character array s, return the number of characters 
-    held inside it.
-*/
-int string_length(char s[])
-{
-    int i = 0;
-    for (; s[i]; i++)
-    {
-    // printf("%c\n", s[i]);
-    // continue;
-    // break;
-    }
-    return i;
-}
-
-/*
-    Don't worry about this funky 'char *' going on here. We'll get to
-    that shortly. In the meantime, write a function that reverses the 
-    order of string s and outputs the reversed string to the input 
-    array rv. The rv array will have enough space for the reversed 
-    string. Don't forget to terminate the reversed string with a null
-    character. Return the rv array.
-*/
-
-char *reverse_string(char rv[], char s[])
-{
-    int length = string_length(s) - 1;
-    int i = 0;
-    
-
-    for (; i <= length; i++)
-    {
-        rv[i] = s[length - i];
-        // printf("%p\n", &rv[i]);
-        if (i == length)
-        {
-            rv[i + 1] = '\0';
-        }
-    }
-    return rv;
-}
+#include <strings.h>
 
 #ifndef TESTING
 int main(void)
